refactor: Split members API in two parts: producer and consumer
The consumer API can be used
when navigating or modifying the tree
once it was built by Griffe and extensions.

The producer API must be used
when building the objects tree,
so by Griffe itself and by extensions.

Using the consumer API while building
is risky because the consumer API
does ~~magic~~ convenient things
and trigger too-early computation
of different things like inherited members.