refactor: Only consider presence/absence for docstrings truthiness, not emptiness of their value
Griffe extensions will often manipulate the parsed sections
of a docstring rather than its raw value. It means that,
often times docstrings will have an empty value
but dynamically inserted sections. In that case a docstring
truthiness should evaluate to true, not false.

We don't want to parse sections to evaluate truthiness
(we tolerate it in extensions, but should avoid it
in Griffe's main logic), so we simply base our evaluation
on the absence or presence of the docstring (none or instance).