refactor: Use final target in alias proxies
The previous comment stating that we should not use
the final target because it could be of another kind
than the first target makes no sense. If the first target
is an alias, then accessing its proxy would access the
next one, until the final target one is reached.
Aliases chains are just that: chains of aliases
with a concrete object at the end. There cannot be
a concrete object in the middle, therefore there
cannot be different kinds in a chain, only aliases
and a single kind at the end.