refactor: Rename loader `load_module` method to `load`
This method always returned the object specified by the path
given as `module` argument. It was not always a module.
So we also renamed the `module` parameter to `objspec`,
and made it positional-only, so we can have a deprecation
period for `module`, keeping it in the signature as
keyword-only. With this, both positional and keyword uses
are kept functional. The latter will emit a deprecation
warning.