feat: Load private sibling modules by default when resolving aliases
Before, external modules were only loaded to resolve aliases if the "resolve external" option/parameter was set to true. It meant that users had to explicitely set "external" to true to resolve aliases from `ast` to `_ast`, or to explicitely "preload" the private modules in mkdocstrings.

But `external=True` is a bit overpowered, and preloading private modules is a bit annoying.

Now, this external parameter accepts a new value, `None`, which also becomes the default value. When external is set to `None`, the loader will automatically load modules to resolve aliases if they are the private sibling of the originating module. It means that users can just load `ast` and be sure that only aliases pointing to `_ast` will be resolved.