fix: Handle partials as functions while inspecting
This doesn't solve the issue that `functools.partial` does not set the `__module__` attribute of the resulting partial to be the module of the wrapped function, which means that the inspector always treats partials as aliases (since it finds out they were defined in the `functools` module).

This issue cannot be fixed on our side, because special-casing partials to decide that they are always defined in the current module would not work for partials imported from elsewhere.

Therefore, devs *must* set the `__module__` attribute of partials themselves.