feat: Pass down agent to extension hooks
Some extensions need to know the user-selected docstring parser and options. The docstring parser and options can only be accessed through the agent.

Other extensions who create objects might want to trigger the rest of the extensions, which can again only be done if we have a reference to the agent.

Issue-312: https://github.com/mkdocstrings/griffe/issues/312