fix: Pre-emptively expand `__all__` values and wildcard imports before firing the `on_package_loaded` event
Extensions hooking the `on_package_loaded` event might trigger computation of resolved bases and inherited members. If a base class is wildcard imported, and wildcard were not expanded, the base class can't be resolved, and inherited members won't be fetched. Since both are cached, that makes them incorrect for the rest of the of the process.

So now we try to expand wildcard imports as much as possible, without loading external packages.

The best solution would be to stop caching resolved bases and inherited members, but I worry of the performance cost. Accessing members of a class using the subscript syntax would compute the resolved bases, MRO and inherited members each time.