fix: Parse Deprecated sections without version info in Google-style docstrings
Griffe requires a version indicator in Google-style Deprecated sections.
If no such indicator is found, then the user is warned, and *then* the
section **is silently dropped**.

[As a proponent of readable docstrings][1], [and apparently in spirit
with the Google style guide][2], I consider it unacceptable to drop
information from a docstring just because it is not properly formatted
as data.

In this patch, I propose parsing the section without an associated
version number instead, and using the section's text as its... well,
text.

[1]: https://github.com/mkdocstrings/python/issues/166#issue-2341762468 '"the source form of the docstring should be readable"'
[2]: https://github.com/google/styleguide/issues/117#issuecomment-203046319  '"Docstrings are meant for human consumption, please keep them readable for us humans. :)"'

Issue-352: https://github.com/mkdocstrings/griffe/issues/352