refactor: Be more consistent regarding not overriding submodules with aliases
Our documentation states that a submodule will always get precedence over an alias of the same name. But the code only avoided replacing a submodule when the alias was pointing to the module.

With this change, we avoid replacing the submodule with *any* alias of the same name, not just aliases that point to the submodule.

The reason we do this is that Griffe suffers from a limitation where a module cannot store both a submodule and anoter member of the same name. We previously decided not to lift this limitation, because we consider this kind of name-shadowing bad practice. We document it in our "Recommendations / Python code" page.